v36: Pinned k3s image versions
Currently the dev environment resolves k3d versions dynamically, so
versions can change under us.

To solve this, this change adds to this repo. a summary of k3s image
metadata, including image digests. `just-k3d` builds an image name from
a logical k8s version in the k3s-images.json file.

To accomodate this, the `tools` image now stores binaries in `/bin` and
stores the k3s image metadata in `/etc`.

The `skopeo` utility is added to the devcontainer for docker image
inspection, as it now used by this repo.