v36: Pin k3s image versions (#16)
Currently the dev environment resolves k3d versions dynamically, so
versions can change under us.

To solve this, this change adds a summary of k3s image metadata,
including image digests, to this repo. `just-k3d` builds an image name
from a logical k8s version in the k3s-channels.json file.

To accomodate this, the `tools` image now stores binaries in `/bin` and
stores the k3s channel metadata in `/etc`.

The `skopeo` utility is added to the devcontainer for docker image
inspection, as it now used by this repo.