fix(sync-k3s-images): skip updating on inconsistent release state
The k3s release index may be updated before container images are available to
use. This change updates the sync-k3s-images recipe to fail if any release images
cannot be found. The CI workflow is updated to allow failures. The workflow fails
if the sync recipe modifies the cached image digests.